% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{Differential}
\alias{Differential}
\title{Differential class}
\usage{
Differential(equation = character(0))
}
\arguments{
\item{equation}{a character

Building basis for defining an ODE}
}
\description{
Differential class
}
\section{Properties}{

\describe{
\item{equation}{The basis for an ODE}
}
}

