% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{print.Differential}
\alias{print.Differential}
\title{print Method for Differential object}
\arguments{
\item{x}{An object of class \code{Differential}}
}
\value{
Invisibly returns \code{x}
}
\description{
Print the equation of the function
}
\examples{
d<-Differential("-3y")
d   #implicit print
print(d)    #explicit print

}
