% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{print.ODEFunction}
\alias{print.ODEFunction}
\title{print Method for ODEFunction object}
\arguments{
\item{x}{An object of class \code{ODEFunction}}
}
\value{
Invisibly returns \code{x}
}
\description{
Print the equation of the ODEFunction
}
\examples{
d<-ODEFunction("-3y")
d   #implicit print
print(d)    #explicit print

}
