% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{Differential}
\alias{Differential}
\title{Differential class}
\usage{
Differential(x)
}
\arguments{
\item{x}{a character}
}
\description{
Differential class
}
\section{Properties}{

\describe{
\item{equation}{The basis for an ODE}
}
}

